requirements.txt
cycler==0.10.0
invoke==0.23.0
kiwisolver==1.0.1
matplotlib==2.2.2
numpy==1.14.3
paho-mqtt==1.3.1
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.4
six==1.11.0
gabby==0.0.5
